Moved the delete method to the api mod
