+++ src/transaction/api.rs
@@ -22,3 +22,22 @@ pub async fn get_transaction_v2(
 
     Ok(HttpResponse::Ok().json(transaction.unwrap()))
 }
+
+pub async fn delete_transaction(
+    params: web::Path<datastruct::IdRequest>,
+    pool: web::Data<Pool<SqliteConnectionManager>>,
+) -> Result<HttpResponse, Error> {
+    let result = db::remove_transaction(pool.get().unwrap(), params.id);
+
+    match result {
+        Ok(_v) => {
+            let result = json!({
+                "status": "DELETED",
+                "id": params.id,
+            });
+
+            Ok(HttpResponse::Ok().json(result))
+        }
+        Err(_e) => Ok(HttpResponse::InternalServerError().finish()),
+    }
+}
